Arduino Main
The main code to be run on the Arduino. It reads all messages and interprets them correctly as well as answers what the outcome is. Guards are put in place so we don't squeeze things too hard, which raise a soft-stop and ultimately a hard-stop, where no further move commands are executed before a reset of the error flags