Added Gripper functions, load settings from config
open_gripper: ready for testing
close_gripper: ready for testing
activate_solenoid: ready for testing
deactivate solenoid: ready for testing
request_info: redy for testing

There was a bug present in the pressureLow flag. Fixed
Added precision variables so that the relative error may be controlled with a relative term plus a bias.
Added "arduino_config" in the config file and added the parameters to the load function.